Update lead generation form to two-step approach with email first
