+++ app/contexts/LeadFormContext.tsx
@@ -1,13 +1,12 @@
 'use client';
 
 import React, { createContext, useState, useContext, ReactNode } from 'react';
-import TypeformPopup from '../components/ui/TypeformPopup';
+import LeadForm from '../components/ui/LeadForm';
 
 type FormData = {
   revenue: string;
   budget: string;
   website: string;
-  instagram: string;
   firstName: string;
   lastName: string;
   phone: string;
@@ -20,6 +19,7 @@ type LeadFormContextType = {
   closeLeadForm: () => void;
   isFormOpen: boolean;
   isFormSubmitted: boolean;
+  capturedEmail: string | null;
 };
 
 const LeadFormContext = createContext<LeadFormContextType | undefined>(undefined);
@@ -39,6 +39,7 @@ type LeadFormProviderProps = {
 export const LeadFormProvider = ({ children }: LeadFormProviderProps) => {
   const [isFormOpen, setIsFormOpen] = useState(false);
   const [isFormSubmitted, setIsFormSubmitted] = useState(false);
+  const [capturedEmail, setCapturedEmail] = useState<string | null>(null);
 
   const openLeadForm = () => {
     setIsFormOpen(true);
@@ -50,8 +51,18 @@ export const LeadFormProvider = ({ children }: LeadFormProviderProps) => {
 
   const handleFormSubmit = (data: FormData) => {
     // Here you would typically send the data to your backend
+    console.log('Form submitted:', data);
+    setCapturedEmail(data.email);
     setIsFormOpen(false);
     setIsFormSubmitted(true);
+    
+    // You can add API calls here to submit the data
+    // Example:
+    // fetch('/api/leads', {
+    //   method: 'POST',
+    //   headers: { 'Content-Type': 'application/json' },
+    //   body: JSON.stringify(data),
+    // });
   };
 
   return (
@@ -61,12 +72,13 @@ export const LeadFormProvider = ({ children }: LeadFormProviderProps) => {
         closeLeadForm,
         isFormOpen,
         isFormSubmitted,
+        capturedEmail,
       }}
     >
       {children}
       
-      {/* Typeform Popup */}
-      <TypeformPopup 
+      {/* Lead Form */}
+      <LeadForm 
         isOpen={isFormOpen} 
         onClose={closeLeadForm} 
         onSubmit={handleFormSubmit}
